Create Create Heart Shape
::before creates a pseudo-element that is the first child of the selected element; ::after creates a pseudo-element that is the last child of the selected element
For the ::before and ::after pseudo-elements to function properly, they must have a defined content property. This property is usually used to add things like a photo or text to the selected element. When the ::before and ::after pseudo-elements are used to make shapes, the content property is still required, but it's set to an empty string.